+++ example/src/pages/ButtonDemo.vue
@@ -1,13 +1,13 @@
 <template>
   <div>
     <Card title="基础用法">
-      <BaseButton large primary round>aaa</BaseButton>
+      <BaseButton primary round>aaa</BaseButton>
     </Card>
 
     <Card title="其他用法">
       <BaseButton success circle>b</BaseButton>
     </Card>
-    <BaseButton large primary :href="'http://www.baidu.com'">cc</BaseButton>
+    <BaseButton large primary>cc</BaseButton>
   </div>
 </template>