+++ packages/button/src/Button.vue
@@ -2,21 +2,7 @@
   <component
     v-bind="{ ...getButtonProps() }"
     :is="tagType"
-    :class="{
-      [styles.base]: true,
-      [styles.primary]: primary,
-      [styles.secondary]: secondary,
-      [styles.success]: success,
-      [styles.danger]: danger,
-      [styles.info]: info,
-      [styles.warning]: warning,
-      [styles.active]: active,
-      [styles.plain]: plain,
-      [styles.small]: small,
-      [styles.large]: large,
-      [styles.round]: round,
-      [styles.disabled]: disableButton,
-    }"
+    :class="btnClass"
     :href="href"
   >
     <slot></slot>
@@ -38,13 +24,15 @@ export default defineComponent({
     tag: { type: [String, Object] },
     href: { type: String },
     primary: { type: Boolean, default: false },
-    secondary: { type: Boolean, default: false },
+    // secondary: { type: Boolean, default: false },
     success: { type: Boolean, default: false },
     danger: { type: Boolean, default: false },
     info: { type: Boolean, default: false },
     warning: { type: Boolean, default: false },
+    pink: { type: Boolean, defult: false },
     active: { type: Boolean, default: false },
     plain: { type: Boolean, default: false },
+    hollow: { type: Boolean, default: false },
     small: { type: Boolean, default: false },
     large: { type: Boolean, default: false },
     round: { type: Boolean, default: false },
@@ -70,30 +58,113 @@ export default defineComponent({
 
     const { getButtonProps } = useBaseButton(userProps, ctx)
 
-    const styles = {
-      base: style`inline-block border no-underline rounded-md px-4 py-2 transition duration-500 ease select-none focus:outline-none hover:shadow-lg text-sm mr-2 whitespace-nowrap`,
-      primary: style`border-brand-500 bg-brand-500 text-white hover:bg-brand-700`,
-      success: style`border-green-500 bg-green-500 text-white hover:bg-green-700`,
-      danger: style`border-red-500 bg-red-500 text-white hover:bg-red-700`,
-      warning: style`border-yellow-500 bg-yellow-500 text-white hover:bg-yellow-700`,
-      info: style`border-gray-500 bg-gray-500 text-white hover:bg-gray-700`,
-      secondary: style`border-gray-500 bg-gray-500 text-gray-300 hover:bg-gray-700`,
-      active: tw`${css`
-        filter: brightness(1.1) contrast(150%);
-      `}`,
-      plain: style`border-0 hover:shadow-none`,
-      small: style`text-xs font-medium`,
-      large: style`text-xl`,
-      round: style`rounded-full`,
-      disabled: style`cursor-not-allowed opacity-50`,
-    }
+    const styles = getStyles()
+
+    const btnClass = computed(() => {
+      const themeString =
+        (props.primary && 'primary') ||
+        (props.pink && 'pink') ||
+        (props.success && 'success') ||
+        (props.danger && 'danger') ||
+        (props.info && 'info') ||
+        (props.warning && 'warning') ||
+        'primary'
+
+      return [
+        styles.themeStyles.base,
+        styles.themeStyles[themeString],
+        props.small ? styles.size.small : '',
+        props.large ? styles.size.large : '',
+        props.hollow
+          ? `${styles.hollow.base} ${styles.hollow[themeString]}`
+          : '',
+        props.plain ? styles.plain[themeString] : '',
+        props.active ? styles.active[themeString] : '',
+        props.disabled
+          ? `${styles.disabled.base} ${styles.disabled[themeString]}`
+          : '',
+        props.round ? styles.size.round : '',
+      ]
+    })
 
     return {
       getButtonProps,
       disableButton,
       tagType,
-      styles,
+      btnClass,
     }
   },
 })
+
+const getStyles = () => {
+  const themeStyles = {
+    base: style`inline-block border-0 no-underline rounded-md py-btn-md-y px-2 transition focus:outline-none duration-500 ease select-none text-base mr-2 whitespace-nowrap text-base text-center ${css`
+      min-width: 3.75rem;
+    `}`,
+    primary: style`bg-brand-primary text-content-white hover:bg-brand-hover active:bg-brand-active`,
+    pink: style`bg-pink-primary text-content-white hover:bg-pink-hover active:bg-pink-active`,
+    success: style`bg-success-primary text-content-white hover:bg-success-hover active:bg-success-active`,
+    danger: style`bg-error-primary text-content-white hover:bg-error-hover active:bg-error-active`,
+    warning: style`bg-warning-primary text-content-white hover:bg-warning-hover active:bg-warning-active`,
+    info: style`bg-info-primary text-content-accent hover:bg-info-hover active:bg-info-active`,
+  }
+  const activeStyles = {
+    primary: style`bg-brand-active`,
+    pink: style`bg-pink-active`,
+    success: style`bg-success-active`,
+    danger: style`bg-error-active`,
+    warning: style`bg-warning-active`,
+    info: style`bg-fill-neutral`,
+    secondary: style`border-brand-active text-brand-active`,
+  }
+
+  const disableStyles = {
+    base: style`cursor-not-allowed`,
+    primary: style`bg-brand-forbid`,
+    pink: style`bg-pink-forbid`,
+    success: style`bg-success-forbid`,
+    danger: style`bg-error-forbid`,
+    warning: style`bg-warning-forbid`,
+    info: style`bg-fill-gray`,
+  }
+
+  const hollowStyles = {
+    base: style`bg-fill-white border`,
+    primary: style`text-brand-primary hover:text-content-white`,
+    pink: style`text-pink-primary hover:text-content-white`,
+    success: style`text-success-primary hover:text-content-white`,
+    danger: style`text-error-primary hover:text-content-white`,
+    warning: style`text-warning-primary hover:text-content-white`,
+    info: style`text-content-accent hover:bg-success-primary hover:text-content-primary`,
+  }
+
+  const plainStyles = {
+    // base: style``,
+    primary: style`bg-brand-light text-brand-primary hover:text-content-white`,
+    pink: style`bg-pink-light text-pink-primary hover:text-content-white`,
+    success: style`bg-success-light text-success-primary hover:text-content-white`,
+    danger: style`bg-error-light text-error-primary hover:text-content-white`,
+    warning: style`bg-warning-light text-warning-primary hover:text-content-white`,
+    info: style`bg-fill-light text-content-accent hover:text-content-primary`,
+    // secondary: style`bg-brand-borbid`,
+  }
+  const sizeStyles = {
+    small: style`text-sm font-medium py-btn-sm-y ${css`
+      min-width: 3.5rem;
+    `}`,
+    large: style`text-base py-btn-lg-y ${css`
+      width: 4rem;
+    `}`,
+    round: style`rounded-full`,
+  }
+
+  return {
+    themeStyles: themeStyles,
+    active: activeStyles,
+    disabled: disableStyles,
+    hollow: hollowStyles,
+    plain: plainStyles,
+    size: sizeStyles,
+  }
+}
 </script>