+++ packages/confirm/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @apathia/apathia.confirm
 
+## 1.0.2
+
+### Patch Changes
+
+- change default theme
+- Updated dependencies
+  - @apathia/apathia.button@1.0.2
+  - @apathia/apathia.modal@1.0.2
+  - @apathia/apathia.twind@1.0.2
+  - @apathia/apathia.custom-render@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes