+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @apathia/apathia.hooks
 
+## 1.0.2
+
+### Patch Changes
+
+- change default theme
+- Updated dependencies
+  - @apathia/apathia.shared@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes