+++ packages/loading/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @apathia/apathia.loading
 
+## 1.0.2
+
+### Patch Changes
+
+- change default theme
+- Updated dependencies
+  - @apathia/apathia.twind@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes