+++ packages/pagination/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @apathia/apathia.pagination
 
+## 1.0.2
+
+### Patch Changes
+
+- change default theme
+- Updated dependencies
+  - @apathia/apathia.twind@1.0.2
+  - @apathia/apathia.shared@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes