+++ packages/upload/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @apathia/apathia.upload
 
+## 1.0.2
+
+### Patch Changes
+
+- change default theme
+- Updated dependencies
+  - @apathia/apathia.alert@1.0.2
+  - @apathia/apathia.button@1.0.2
+  - @apathia/apathia.input@1.0.2
+  - @apathia/apathia.twind@1.0.2
+  - @apathia/apathia.hooks@1.0.2
+  - @apathia/apathia.icon@1.0.2
+  - @apathia/apathia.shared@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes